fix: clippy warnings and Windows test client TCP support (#112)
- Fix empty line after doc comment in ipc.rs
- Remove unnecessary borrow in daemon_lifecycle.rs
- Remove unnecessary u32 cast in runner/mod.rs
- Remove explicit deref in shell.rs test
- Add TCP-based DaemonClient for Windows tests (mock daemon uses TCP)
- Extract shared send_request() to deduplicate request/raw_request

+++ crates/daemon/src/runner/mod.rs
@@ -475,7 +475,7 @@ impl RunnerManager {
             let (state, pid) = if is_service {
                 // Service runners survive daemon restart — always check if process is alive
                 match find_runner_pid(&entry.config.work_dir).await {
-                    Some(pid) => (RunnerState::Online, Some(pid as u32)),
+                    Some(pid) => (RunnerState::Online, Some(pid)),
                     None => {
                         // Service runner's process died — should be restarted
                         if entry.was_running {
